refactor: align sqlite adapter boundaries
