+++ src/infra/adapters/sqlite/mod.rs
@@ -1,24 +1,31 @@
-use std::borrow::Cow;
 use std::collections::{HashMap, HashSet};
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::time::Instant;
 
 use async_trait::async_trait;
 use serde::Deserialize;
 
 use crate::app::ports::outbound::{DbOperationError, MetadataProvider, QueryExecutor};
+#[cfg(test)]
+use crate::domain::QueryValue;
 use crate::domain::{
     Column, ColumnAttributes, CommandTag, DatabaseMetadata, FkAction, ForeignKey, Index,
-    IndexAttributes, IndexType, QueryResult, QuerySource, QueryValue, Schema, Table,
-    TableSignature, TableSummary, Trigger, TriggerEvent, TriggerTiming, UNRESOLVED_FK_COLUMN,
-    WriteExecutionResult,
+    IndexAttributes, IndexType, QueryResult, QuerySource, Schema, Table, TableSignature,
+    TableSummary, UNRESOLVED_FK_COLUMN, WriteExecutionResult,
 };
 
-mod cli;
-mod error;
 mod sql;
+mod sqlite3;
 
-use cli::SqliteCli;
+#[cfg(test)]
+use sqlite3::BUSY_TIMEOUT_MS;
+use sqlite3::SqliteCli;
+use sqlite3::parser::{
+    aggregate_sqlite_command_tag, append_changes_query, command_tag_result,
+    is_sqlite_rerunnable_export_query, last_sqlite_result_set, parse_affected_rows,
+    parse_count_result, quoted_to_query_result, sqlite_adhoc_execution_query,
+    sqlite_export_not_rerunnable_error, sqlite_probe_marker, sqlite_statement_tags,
+    statement_counts_as_select_tag, strip_sqlite_probes, try_split_sqlite_statements,
+};
 
 const MAIN_SCHEMA: &str = "main";
 
@@ -81,12 +88,6 @@ struct RawForeignKey {
     to: Option<String>,
     on_update: String,
     on_delete: String,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-struct RawTrigger {
-    name: String,
-    sql: Option<String>,
 }
 
 #[derive(Debug, Clone, Deserialize)]
@@ -406,23 +407,6 @@ impl SqliteAdapter {
             .and_then(|row| row.sql)
     }
 
-    async fn triggers(&self, path: &str, table: &str) -> Result<Vec<Trigger>, DbOperationError> {
-        let raw: Vec<RawTrigger> = self
-            .cli
-            .execute_json(path, &sql::trigger_list_query(table))
-            .await?;
-        let mut triggers = Vec::new();
-
-        for raw in raw {
-            if let Some(sql) = raw.sql {
-                triggers.push(parse_sqlite_trigger(&raw.name, &sql)?);
-            }
-        }
-
-        triggers.sort_by(|a, b| a.name.cmp(&b.name));
-        Ok(triggers)
-    }
-
     async fn foreign_keys(
         &self,
         path: &str,
@@ -629,7 +613,7 @@ impl SqliteAdapter {
             foreign_keys: self.foreign_keys(path, table).await?,
             indexes,
             rls: None,
-            triggers: self.triggers(path, table).await?,
+            triggers: Vec::new(),
             row_count_estimate: if include_row_count {
                 self.row_count(path, table).await
             } else {
@@ -688,20 +672,6 @@ impl SqliteAdapter {
                 fk.reference_resolved
             )
         }));
-        parts.extend(detail.triggers.iter().map(|trigger| {
-            format!(
-                "trg={}:{}:{}:{}",
-                trigger.name,
-                trigger.timing,
-                trigger
-                    .events
-                    .iter()
-                    .map(ToString::to_string)
-                    .collect::<Vec<_>>()
-                    .join(","),
-                trigger.function_name
-            )
-        }));
 
         Ok(TableSignature {
             schema: MAIN_SCHEMA.to_string(),
@@ -709,1349 +679,6 @@ impl SqliteAdapter {
             signature: parts.join("|"),
         })
     }
-}
-
-fn is_ident_char(byte: u8) -> bool {
-    byte.is_ascii_alphanumeric() || byte == b'_'
-}
-
-fn skip_quoted(bytes: &[u8], mut i: usize, quote: u8) -> usize {
-    i += 1;
-    while i < bytes.len() {
-        if bytes[i] == quote {
-            if i + 1 < bytes.len() && bytes[i + 1] == quote {
-                i += 2;
-            } else {
-                return i + 1;
-            }
-        } else {
-            i += 1;
-        }
-    }
-    i
-}
-
-fn skip_bracket_quoted(bytes: &[u8], mut i: usize) -> usize {
-    i += 1;
-    while i < bytes.len() {
-        if bytes[i] == b']' {
-            return i + 1;
-        }
-        i += 1;
-    }
-    i
-}
-
-/// Returns the next SQL keyword and the byte offset immediately after it.
-fn next_keyword_from(sql: &str, mut i: usize) -> Option<(&str, usize)> {
-    let bytes = sql.as_bytes();
-    while i < bytes.len() {
-        match bytes[i] {
-            b'-' if i + 1 < bytes.len() && bytes[i + 1] == b'-' => {
-                while i < bytes.len() && bytes[i] != b'\n' {
-                    i += 1;
-                }
-            }
-            b'/' if i + 1 < bytes.len() && bytes[i + 1] == b'*' => {
-                i += 2;
-                while i + 1 < bytes.len() && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                    i += 1;
-                }
-                if i + 1 < bytes.len() {
-                    i += 2;
-                }
-            }
-            b'\'' | b'"' | b'`' => {
-                i = skip_quoted(bytes, i, bytes[i]);
-            }
-            b'[' => {
-                i = skip_bracket_quoted(bytes, i);
-            }
-            b if b.is_ascii_alphabetic() => {
-                let start = i;
-                while i < bytes.len() && is_ident_char(bytes[i]) {
-                    i += 1;
-                }
-                return Some((&sql[start..i], i));
-            }
-            _ => i += 1,
-        }
-    }
-    None
-}
-
-fn first_keyword(sql: &str) -> &str {
-    next_keyword_from(sql, 0).map_or("", |(keyword, _)| keyword)
-}
-
-fn second_keyword(sql: &str) -> Option<&str> {
-    let (_, end) = next_keyword_from(sql, 0)?;
-    next_keyword_from(sql, end).map(|(keyword, _)| keyword)
-}
-
-fn contains_keyword(sql: &str, expected: &str) -> bool {
-    let mut offset = 0;
-    while let Some((keyword, end)) = next_keyword_from(sql, offset) {
-        if keyword.eq_ignore_ascii_case(expected) {
-            return true;
-        }
-        offset = end;
-    }
-    false
-}
-
-fn sqlite_trigger_parse_error(sql: &str, detail: &str) -> DbOperationError {
-    DbOperationError::MetadataParseFailed(format!(
-        "sqlite trigger parse failed ({detail}): {}",
-        sql.chars().take(120).collect::<String>()
-    ))
-}
-
-fn skip_optional_if_not_exists(sql: &str, pos: usize) -> usize {
-    let Some((keyword, next)) = next_keyword_from(sql, pos) else {
-        return pos;
-    };
-    if !keyword.eq_ignore_ascii_case("IF") {
-        return pos;
-    }
-    let Some((not, next)) = next_keyword_from(sql, next) else {
-        return pos;
-    };
-    if !not.eq_ignore_ascii_case("NOT") {
-        return pos;
-    }
-    let Some((exists, next)) = next_keyword_from(sql, next) else {
-        return pos;
-    };
-    if exists.eq_ignore_ascii_case("EXISTS") {
-        next
-    } else {
-        pos
-    }
-}
-
-fn skip_object_reference(sql: &str, pos: usize) -> usize {
-    let bytes = sql.as_bytes();
-    let mut i = pos;
-    while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-        i += 1;
-    }
-    let start = match bytes.get(i) {
-        Some(b'"' | b'\'' | b'`') => skip_quoted(bytes, i, bytes[i]),
-        Some(b'[') => skip_bracket_quoted(bytes, i),
-        Some(b) if b.is_ascii_alphanumeric() || *b == b'_' => {
-            let Some((_, end)) = next_keyword_from(sql, i) else {
-                return i;
-            };
-            end
-        }
-        _ => return i,
-    };
-    if bytes.get(start) == Some(&b'.')
-        && let Some((_, end)) = next_keyword_from(sql, start + 1)
-    {
-        return end;
-    }
-    start
-}
-
-fn skip_update_of_clause(sql: &str, pos: usize) -> usize {
-    let Some((keyword, next)) = next_keyword_from(sql, pos) else {
-        return pos;
-    };
-    if !keyword.eq_ignore_ascii_case("OF") {
-        return pos;
-    }
-
-    let mut pos = next;
-    loop {
-        let Some((keyword, next)) = next_keyword_from(sql, pos) else {
-            return pos;
-        };
-        match keyword.to_ascii_uppercase().as_str() {
-            "INSERT" | "UPDATE" | "DELETE" | "ON" | "FOR" | "WHEN" | "BEGIN" => return pos,
-            _ => pos = next,
-        }
-    }
-}
-
-fn parse_sqlite_trigger_events(
-    sql: &str,
-    pos: usize,
-) -> Result<(Vec<TriggerEvent>, usize), DbOperationError> {
-    let mut events = Vec::new();
-    let mut pos = pos;
-    loop {
-        let Some((keyword, next)) = next_keyword_from(sql, pos) else {
-            return Err(sqlite_trigger_parse_error(sql, "missing trigger event"));
-        };
-        if keyword.eq_ignore_ascii_case("ON") {
-            break;
-        }
-
-        match keyword.to_ascii_uppercase().as_str() {
-            "INSERT" => events.push(TriggerEvent::Insert),
-            "UPDATE" => {
-                events.push(TriggerEvent::Update);
-                pos = skip_update_of_clause(sql, next);
-                continue;
-            }
-            "DELETE" => events.push(TriggerEvent::Delete),
-            _ => return Err(sqlite_trigger_parse_error(sql, "unsupported trigger event")),
-        }
-        pos = next;
-    }
-
-    if events.is_empty() {
-        return Err(sqlite_trigger_parse_error(sql, "no trigger events"));
-    }
-
-    Ok((events, pos))
-}
-
-fn parse_sqlite_trigger_header(
-    sql: &str,
-    pos: usize,
-) -> Result<(TriggerTiming, Vec<TriggerEvent>, usize), DbOperationError> {
-    let Some((keyword, next)) = next_keyword_from(sql, pos) else {
-        return Err(sqlite_trigger_parse_error(
-            sql,
-            "missing trigger timing or event",
-        ));
-    };
-
-    match keyword.to_ascii_uppercase().as_str() {
-        "BEFORE" => {
-            let (events, pos) = parse_sqlite_trigger_events(sql, next)?;
-            Ok((TriggerTiming::Before, events, pos))
-        }
-        "AFTER" => {
-            let (events, pos) = parse_sqlite_trigger_events(sql, next)?;
-            Ok((TriggerTiming::After, events, pos))
-        }
-        "INSTEAD" => {
-            let Some((of, next)) = next_keyword_from(sql, next) else {
-                return Err(sqlite_trigger_parse_error(sql, "incomplete INSTEAD OF"));
-            };
-            if !of.eq_ignore_ascii_case("OF") {
-                return Err(sqlite_trigger_parse_error(sql, "expected OF after INSTEAD"));
-            }
-            let (events, pos) = parse_sqlite_trigger_events(sql, next)?;
-            Ok((TriggerTiming::InsteadOf, events, pos))
-        }
-        "INSERT" | "UPDATE" | "DELETE" => {
-            let (events, pos) = parse_sqlite_trigger_events(sql, pos)?;
-            Ok((TriggerTiming::Before, events, pos))
-        }
-        _ => Err(sqlite_trigger_parse_error(
-            sql,
-            "unsupported trigger timing or event",
-        )),
-    }
-}
-
-fn parse_sqlite_trigger(trigger_name: &str, sql: &str) -> Result<Trigger, DbOperationError> {
-    let Some((first, pos)) = next_keyword_from(sql, 0) else {
-        return Err(sqlite_trigger_parse_error(sql, "missing CREATE"));
-    };
-    if !first.eq_ignore_ascii_case("CREATE") {
-        return Err(sqlite_trigger_parse_error(sql, "expected CREATE"));
-    }
-    let Some((second, pos)) = next_keyword_from(sql, pos) else {
-        return Err(sqlite_trigger_parse_error(sql, "missing TRIGGER"));
-    };
-    if !second.eq_ignore_ascii_case("TRIGGER") {
-        return Err(sqlite_trigger_parse_error(sql, "expected TRIGGER"));
-    }
-
-    let mut pos = pos;
-    if let Some((keyword, next)) = next_keyword_from(sql, pos)
-        && keyword.eq_ignore_ascii_case("TEMP")
-    {
-        pos = next;
-    }
-    pos = skip_optional_if_not_exists(sql, pos);
-    pos = skip_object_reference(sql, pos);
-
-    let (timing, events, _) = parse_sqlite_trigger_header(sql, pos)?;
-
-    Ok(Trigger {
-        name: trigger_name.to_string(),
-        timing,
-        events,
-        function_name: sql.to_string(),
-        security_definer: false,
-    })
-}
-
-fn is_create_trigger_prefix(sql: &str) -> bool {
-    let Some((first, pos)) = next_keyword_from(sql, 0) else {
-        return false;
-    };
-    if !first.eq_ignore_ascii_case("CREATE") {
-        return false;
-    }
-    let Some((second, pos)) = next_keyword_from(sql, pos) else {
-        return false;
-    };
-    if second.eq_ignore_ascii_case("TEMP") || second.eq_ignore_ascii_case("TEMPORARY") {
-        let Some((third, _)) = next_keyword_from(sql, pos) else {
-            return false;
-        };
-        return third.eq_ignore_ascii_case("TRIGGER");
-    }
-    second.eq_ignore_ascii_case("TRIGGER")
-}
-
-fn is_dotted_identifier_suffix(sql: &str, keyword_start: usize) -> bool {
-    let mut index = keyword_start;
-    while index > 0 {
-        index -= 1;
-        match sql.as_bytes()[index] {
-            byte if byte.is_ascii_whitespace() => {}
-            b'.' => return true,
-            _ => return false,
-        }
-    }
-    false
-}
-
-fn is_sqlite_trigger_body_end(
-    keyword: &str,
-    in_trigger_body: bool,
-    trigger_body_stmt_start: bool,
-    sql: &str,
-    keyword_start: usize,
-) -> bool {
-    in_trigger_body
-        && trigger_body_stmt_start
-        && keyword.eq_ignore_ascii_case("END")
-        && !is_dotted_identifier_suffix(sql, keyword_start)
-}
-
-fn try_split_sqlite_statements(sql: &str) -> Result<Vec<&str>, DbOperationError> {
-    let bytes = sql.as_bytes();
-    let mut statements = Vec::new();
-    let mut start = 0;
-    let mut i = 0;
-    let mut in_trigger_body = false;
-    let mut trigger_body_stmt_start = false;
-
-    while i < bytes.len() {
-        match bytes[i] {
-            b'-' if i + 1 < bytes.len() && bytes[i + 1] == b'-' => {
-                while i < bytes.len() && bytes[i] != b'\n' {
-                    i += 1;
-                }
-            }
-            b'/' if i + 1 < bytes.len() && bytes[i + 1] == b'*' => {
-                i += 2;
-                while i + 1 < bytes.len() && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                    i += 1;
-                }
-                if i + 1 < bytes.len() {
-                    i += 2;
-                }
-            }
-            b'\'' | b'"' | b'`' => {
-                i = skip_quoted(bytes, i, bytes[i]);
-            }
-            b'[' => {
-                i = skip_bracket_quoted(bytes, i);
-            }
-            b';' if in_trigger_body => {
-                trigger_body_stmt_start = true;
-                i += 1;
-            }
-            b';' => {
-                let statement = sql[start..i].trim();
-                if !statement.is_empty() {
-                    statements.push(statement);
-                }
-                i += 1;
-                start = i;
-                in_trigger_body = false;
-                trigger_body_stmt_start = false;
-            }
-            _ => {
-                if bytes[i].is_ascii_alphabetic()
-                    && is_create_trigger_prefix(&sql[start..])
-                    && let Some((keyword, kw_end)) = next_keyword_from(sql, i)
-                {
-                    if keyword.eq_ignore_ascii_case("BEGIN") {
-                        if in_trigger_body {
-                            trigger_body_stmt_start = false;
-                        } else {
-                            in_trigger_body = true;
-                            trigger_body_stmt_start = true;
-                        }
-                    } else if is_sqlite_trigger_body_end(
-                        keyword,
-                        in_trigger_body,
-                        trigger_body_stmt_start,
-                        sql,
-                        i,
-                    ) {
-                        in_trigger_body = false;
-                        trigger_body_stmt_start = false;
-                    } else if in_trigger_body {
-                        trigger_body_stmt_start = false;
-                    }
-                    i = kw_end;
-                    continue;
-                }
-                i += 1;
-            }
-        }
-    }
-
-    let tail = sql[start..].trim();
-    if !tail.is_empty() {
-        if in_trigger_body {
-            return Err(DbOperationError::QueryFailed(
-                "Unclosed CREATE TRIGGER body".to_string(),
-            ));
-        }
-        if is_create_trigger_prefix(tail) && !contains_keyword(tail, "BEGIN") {
-            return Err(DbOperationError::QueryFailed(
-                "Incomplete CREATE TRIGGER statement".to_string(),
-            ));
-        }
-        statements.push(tail);
-    }
-
-    Ok(statements)
-}
-
-fn is_transaction_control(statement: &str) -> bool {
-    matches!(
-        first_keyword(statement).to_ascii_uppercase().as_str(),
-        "BEGIN" | "COMMIT" | "END" | "ROLLBACK" | "SAVEPOINT" | "RELEASE"
-    )
-}
-
-fn is_write_statement(statement: &str) -> bool {
-    if is_dml_statement(statement) {
-        return true;
-    }
-    matches!(
-        first_keyword(statement).to_ascii_uppercase().as_str(),
-        "CREATE" | "ALTER" | "DROP" | "TRUNCATE"
-    )
-}
-
-fn is_sqlite_rerunnable_export_query(query: &str) -> Result<bool, DbOperationError> {
-    let statements = try_split_sqlite_statements(query)?;
-    Ok(statements.len() == 1
-        && statements
-            .iter()
-            .all(|statement| is_sqlite_rerunnable_export_statement(statement)))
-}
-
-fn is_sqlite_rerunnable_export_statement(statement: &str) -> bool {
-    if is_write_statement(statement)
-        || is_dml_statement(statement)
-        || is_transaction_control(statement)
-    {
-        return false;
-    }
-    match first_keyword(statement).to_ascii_uppercase().as_str() {
-        "SELECT" | "EXPLAIN" | "VALUES" => true,
-        "WITH" => !is_dml_statement(statement),
-        "PRAGMA" => is_read_only_sqlite_export_pragma(statement),
-        _ => false,
-    }
-}
-
-fn sqlite_pragma_name(statement: &str) -> Option<String> {
-    let (_, pragma_end) = next_keyword_from(statement, 0)?;
-    let tail = statement.get(pragma_end..)?.trim_start();
-    let name: String = tail
-        .chars()
-        .take_while(|ch| ch.is_ascii_alphanumeric() || *ch == '_' || *ch == '.')
-        .collect();
-    if name.is_empty() {
-        None
-    } else {
-        Some(
-            name.rsplit('.')
-                .next()
-                .unwrap_or(name.as_str())
-                .to_ascii_lowercase(),
-        )
-    }
-}
-
-fn is_read_only_parameterized_pragma(name: &str) -> bool {
-    matches!(
-        name,
-        "table_info"
-            | "table_xinfo"
-            | "index_info"
-            | "index_xinfo"
-            | "index_list"
-            | "foreign_key_list"
-            | "database_list"
-            | "table_list"
-            | "pragma_list"
-            | "function_list"
-            | "module_list"
-            | "collation_list"
-            | "integrity_check"
-            | "quick_check"
-            | "column_info"
-    )
-}
-
-fn is_read_only_sqlite_export_pragma(statement: &str) -> bool {
-    let Some(name) = sqlite_pragma_name(statement) else {
-        return false;
-    };
-    let has_assignment = statement.contains('=');
-    let has_parenthesized_value = statement.contains('(');
-    let side_effect_without_assignment = (has_parenthesized_value
-        && !is_read_only_parameterized_pragma(&name))
-        || matches!(
-            name.as_str(),
-            "optimize" | "incremental_vacuum" | "wal_checkpoint"
-        );
-    !has_assignment && !side_effect_without_assignment
-}
-
-fn sqlite_export_not_rerunnable_error() -> DbOperationError {
-    DbOperationError::UnsupportedOperation(
-        "Cannot re-execute this query for CSV export because it contains write or DDL statements"
-            .to_string(),
-    )
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum SqliteWrapMode {
-    None,
-    BeginCommit,
-}
-
-fn needs_write_batch_wrap(statements: &[&str]) -> bool {
-    statements.len() > 1 && statements.iter().any(|stmt| is_write_statement(stmt))
-}
-
-fn rollback_has_to_clause(statement: &str) -> bool {
-    if !first_keyword(statement).eq_ignore_ascii_case("ROLLBACK") {
-        return false;
-    }
-    let mut offset = 0;
-    while let Some((keyword, end)) = next_keyword_from(statement, offset) {
-        if keyword.eq_ignore_ascii_case("TO") {
-            return true;
-        }
-        offset = end;
-    }
-    false
-}
-
-fn rollback_to_target(statement: &str) -> Option<&str> {
-    let (_, first_end) = next_keyword_from(statement, 0)?;
-    if !first_keyword(statement).eq_ignore_ascii_case("ROLLBACK") {
-        return None;
-    }
-    let (second, second_end) = next_keyword_from(statement, first_end)?;
-    if second.eq_ignore_ascii_case("TRANSACTION") {
-        let (third, third_end) = next_keyword_from(statement, second_end)?;
-        if !third.eq_ignore_ascii_case("TO") {
-            return None;
-        }
-        let (fourth, fourth_end) = next_keyword_from(statement, third_end)?;
-        if fourth.eq_ignore_ascii_case("SAVEPOINT") {
-            next_keyword_from(statement, fourth_end).map(|(name, _)| name)
-        } else {
-            Some(fourth)
-        }
-    } else if second.eq_ignore_ascii_case("TO") {
-        let (third, third_end) = next_keyword_from(statement, second_end)?;
-        if third.eq_ignore_ascii_case("SAVEPOINT") {
-            next_keyword_from(statement, third_end).map(|(name, _)| name)
-        } else {
-            Some(third)
-        }
-    } else {
-        None
-    }
-}
-
-fn is_rollback_to(statement: &str) -> bool {
-    rollback_to_target(statement).is_some() || rollback_has_to_clause(statement)
-}
-
-fn sqlite_wrap_mode(query: &str) -> Result<SqliteWrapMode, DbOperationError> {
-    let statements = try_split_sqlite_statements(query)?;
-    if !needs_write_batch_wrap(&statements) {
-        return Ok(SqliteWrapMode::None);
-    }
-
-    if statements.iter().any(|stmt| is_transaction_control(stmt)) {
-        return Ok(SqliteWrapMode::None);
-    }
-
-    Ok(SqliteWrapMode::BeginCommit)
-}
-
-fn sqlite_transaction_block(query: &str) -> String {
-    let trimmed = query.trim_end().trim_end_matches(';').trim_end();
-    format!("BEGIN;\n{trimmed}\n;\nCOMMIT")
-}
-
-fn sqlite_execution_query(query: &str) -> Result<Cow<'_, str>, DbOperationError> {
-    Ok(match sqlite_wrap_mode(query)? {
-        SqliteWrapMode::BeginCommit => Cow::Owned(sqlite_transaction_block(query)),
-        SqliteWrapMode::None => Cow::Borrowed(query),
-    })
-}
-
-fn sqlite_probe_marker() -> String {
-    static SEQ: AtomicU64 = AtomicU64::new(0);
-    let nanos = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .map_or(0, |d| d.as_nanos());
-    format!(
-        "__sabiql_sqlite_probe_{}_{}_{}",
-        std::process::id(),
-        nanos,
-        SEQ.fetch_add(1, Ordering::Relaxed)
-    )
-}
-
-fn sqlite_probe_columns(marker: &str) -> (String, String) {
-    (format!("{marker}_stmt"), format!("{marker}_changes"))
-}
-
-fn sqlite_changes_probe(marker: &str, index: usize) -> String {
-    let (stmt_col, changes_col) = sqlite_probe_columns(marker);
-    format!("SELECT {index} AS \"{stmt_col}\", changes() AS \"{changes_col}\"")
-}
-
-fn sqlite_result_probe_columns(marker: &str) -> (String, String) {
-    (
-        format!("{marker}_result_stmt"),
-        format!("{marker}_result_marker"),
-    )
-}
-
-fn sqlite_result_probe(marker: &str, index: usize) -> String {
-    let (stmt_col, marker_col) = sqlite_result_probe_columns(marker);
-    format!("SELECT {index} AS \"{stmt_col}\", '{marker}' AS \"{marker_col}\"")
-}
-
-fn sqlite_adhoc_execution_query(query: &str, marker: &str) -> Result<String, DbOperationError> {
-    let statements = try_split_sqlite_statements(query)?;
-    if statements.is_empty() {
-        return Ok(query.to_string());
-    }
-
-    let wrap_mode = sqlite_wrap_mode(query)?;
-    let mut parts = Vec::with_capacity(statements.len() * 2 + 2);
-    if matches!(wrap_mode, SqliteWrapMode::BeginCommit) {
-        parts.push("BEGIN".to_string());
-    }
-    for (index, statement) in statements.iter().enumerate() {
-        parts.push((*statement).to_string());
-        if is_dml_statement(statement) {
-            parts.push(sqlite_changes_probe(marker, index));
-        }
-        if statement_emits_result_set(statement) {
-            parts.push(sqlite_result_probe(marker, index));
-        }
-    }
-    if matches!(wrap_mode, SqliteWrapMode::BeginCommit) {
-        parts.push("COMMIT".to_string());
-    }
-    Ok(parts.join("\n;\n"))
-}
-
-fn append_changes_query(query: &str) -> Result<String, DbOperationError> {
-    let body = sqlite_execution_query(query)?.trim_end().to_string();
-    // The standalone separator also terminates a trailing line comment before
-    // appending the changes() probe.
-    Ok(format!("{body}\n;\nSELECT changes() AS affected_rows;"))
-}
-
-fn dml_keyword(statement: &str) -> Option<&'static str> {
-    let keyword = first_keyword(statement);
-    if keyword.eq_ignore_ascii_case("INSERT") {
-        return Some("INSERT");
-    }
-    if keyword.eq_ignore_ascii_case("REPLACE") {
-        return Some("INSERT");
-    }
-    if keyword.eq_ignore_ascii_case("UPDATE") {
-        return Some("UPDATE");
-    }
-    if keyword.eq_ignore_ascii_case("DELETE") {
-        return Some("DELETE");
-    }
-    if !keyword.eq_ignore_ascii_case("WITH") {
-        return None;
-    }
-
-    let mut offset = 0;
-    while let Some((keyword, end)) = next_keyword_from(statement, offset) {
-        if keyword.eq_ignore_ascii_case("INSERT") {
-            return Some("INSERT");
-        }
-        if keyword.eq_ignore_ascii_case("REPLACE") {
-            return Some("INSERT");
-        }
-        if keyword.eq_ignore_ascii_case("UPDATE") {
-            return Some("UPDATE");
-        }
-        if keyword.eq_ignore_ascii_case("DELETE") {
-            return Some("DELETE");
-        }
-        offset = end;
-    }
-    None
-}
-
-fn is_dml_statement(statement: &str) -> bool {
-    dml_keyword(statement).is_some()
-}
-
-fn statement_emits_result_set(statement: &str) -> bool {
-    let keyword = first_keyword(statement);
-    if keyword.eq_ignore_ascii_case("SELECT")
-        || keyword.eq_ignore_ascii_case("PRAGMA")
-        || keyword.eq_ignore_ascii_case("EXPLAIN")
-        || keyword.eq_ignore_ascii_case("VALUES")
-    {
-        return true;
-    }
-    if is_dml_statement(statement) {
-        return contains_keyword(statement, "RETURNING");
-    }
-    keyword.eq_ignore_ascii_case("WITH")
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct QuotedRecord {
-    offset: usize,
-    values: Vec<QueryValue>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct SplitSegment {
-    offset: usize,
-    text: String,
-}
-
-fn split_outside_sqlite_quotes(
-    input: &str,
-    delimiter: u8,
-) -> Result<Vec<SplitSegment>, DbOperationError> {
-    let mut segments = Vec::new();
-    let mut start = 0usize;
-    let mut in_quote = false;
-    let bytes = input.as_bytes();
-    let mut i = 0usize;
-    while i < bytes.len() {
-        match bytes[i] {
-            b'\'' if in_quote && i + 1 < bytes.len() && bytes[i + 1] == b'\'' => {
-                i += 2;
-            }
-            b'\'' => {
-                in_quote = !in_quote;
-                i += 1;
-            }
-            byte if byte == delimiter && !in_quote => {
-                segments.push(SplitSegment {
-                    offset: start,
-                    text: input[start..i].to_string(),
-                });
-                start = i + 1;
-                i += 1;
-            }
-            _ => i += 1,
-        }
-    }
-    if in_quote {
-        return Err(DbOperationError::MetadataParseFailed(
-            "unterminated SQLite quoted output".to_string(),
-        ));
-    }
-    segments.push(SplitSegment {
-        offset: start,
-        text: input[start..].to_string(),
-    });
-    Ok(segments)
-}
-
-fn split_quoted_records(stdout: &str) -> Result<Vec<SplitSegment>, DbOperationError> {
-    let mut records = split_outside_sqlite_quotes(stdout, b'\n')?
-        .into_iter()
-        .map(|segment| SplitSegment {
-            offset: segment.offset,
-            text: segment.text.trim_end_matches('\r').to_string(),
-        })
-        .collect::<Vec<_>>();
-    records.retain(|segment| !segment.text.is_empty());
-    Ok(records)
-}
-
-fn split_quoted_fields(record: &str) -> Result<Vec<String>, DbOperationError> {
-    split_outside_sqlite_quotes(record, b',')
-        .map(|segments| segments.into_iter().map(|segment| segment.text).collect())
-}
-
-fn unquote_sql_text(value: &str) -> String {
-    value[1..value.len() - 1].replace("''", "'")
-}
-
-fn decode_hex_text(hex: &str) -> Result<String, DbOperationError> {
-    let bytes = decode_hex_bytes(hex)?;
-    Ok(String::from_utf8_lossy(&bytes).into_owned())
-}
-
-fn parse_unistr_inner_sql_escapes(value: &str) -> Result<String, DbOperationError> {
-    let inner = value
-        .strip_prefix("unistr(")
-        .and_then(|rest| rest.strip_suffix(')'))
-        .ok_or_else(|| {
-            DbOperationError::MetadataParseFailed("invalid SQLite unistr literal".to_string())
-        })?;
-    let inner = inner
-        .strip_prefix('\'')
-        .and_then(|rest| rest.strip_suffix('\''))
-        .ok_or_else(|| {
-            DbOperationError::MetadataParseFailed("invalid SQLite unistr literal".to_string())
-        })?;
-
-    let mut decoded = String::new();
-    let mut chars = inner.chars();
-    while let Some(ch) = chars.next() {
-        match ch {
-            '\'' => {
-                let next = chars.next().ok_or_else(|| {
-                    DbOperationError::MetadataParseFailed(
-                        "invalid SQLite unistr SQL string quote".to_string(),
-                    )
-                })?;
-                if next != '\'' {
-                    return Err(DbOperationError::MetadataParseFailed(
-                        "invalid SQLite unistr SQL string quote".to_string(),
-                    ));
-                }
-                decoded.push('\'');
-            }
-            '\\' => {
-                let next = chars.next().ok_or_else(|| {
-                    DbOperationError::MetadataParseFailed(
-                        "invalid SQLite unistr escape sequence".to_string(),
-                    )
-                })?;
-                if next == '\\' {
-                    decoded.push('\\');
-                } else {
-                    decoded.push('\\');
-                    decoded.push(next);
-                }
-            }
-            ch => decoded.push(ch),
-        }
-    }
-    Ok(decoded)
-}
-
-fn decode_sqlite_nul_text_transport(text: &str) -> Result<Option<String>, DbOperationError> {
-    if let Some(hex) = text.strip_prefix(&sql::sqlite_nul_text_sentinel()) {
-        return decode_hex_text(hex).map(Some);
-    }
-    if let Some(hex) = text.strip_prefix(sql::PREVIEW_TRANSPORT_UNISTR_PREFIX) {
-        return decode_hex_text(hex).map(Some);
-    }
-    Ok(None)
-}
-
-fn decode_preview_transport_unistr(value: &str) -> Result<Option<String>, DbOperationError> {
-    let inner = parse_unistr_inner_sql_escapes(value)?;
-    decode_sqlite_nul_text_transport(&inner)
-}
-
-fn decode_hex_bytes(hex: &str) -> Result<Vec<u8>, DbOperationError> {
-    if !hex.len().is_multiple_of(2) {
-        return Err(DbOperationError::MetadataParseFailed(
-            "invalid SQLite BLOB hex literal".to_string(),
-        ));
-    }
-    let mut bytes = Vec::with_capacity(hex.len() / 2);
-    let mut chars = hex.as_bytes().chunks_exact(2);
-    for pair in &mut chars {
-        let raw = std::str::from_utf8(pair)
-            .map_err(|error| DbOperationError::MetadataParseFailed(error.to_string()))?;
-        let byte = u8::from_str_radix(raw, 16)
-            .map_err(|error| DbOperationError::MetadataParseFailed(error.to_string()))?;
-        bytes.push(byte);
-    }
-    Ok(bytes)
-}
-
-fn parse_quoted_value(
-    value: &str,
-    source: QuerySource,
-    decode_preview_transport: bool,
-) -> Result<QueryValue, DbOperationError> {
-    if value == "NULL" {
-        return Ok(QueryValue::Null);
-    }
-    if value.starts_with('\'') && value.ends_with('\'') && value.len() >= 2 {
-        let text = unquote_sql_text(value);
-        if source == QuerySource::Preview
-            && decode_preview_transport
-            && let Some(decoded) = decode_sqlite_nul_text_transport(&text)?
-        {
-            return Ok(QueryValue::Text(decoded));
-        }
-        return Ok(QueryValue::Text(text));
-    }
-    if value.starts_with("unistr(") && value.ends_with(')') {
-        if source == QuerySource::Preview
-            && decode_preview_transport
-            && let Some(text) = decode_preview_transport_unistr(value)?
-        {
-            return Ok(QueryValue::Text(text));
-        }
-        return Ok(QueryValue::SqlLiteral(value.to_string()));
-    }
-    if value.len() >= 3
-        && value.as_bytes()[1] == b'\''
-        && value.ends_with('\'')
-        && value.as_bytes()[0].eq_ignore_ascii_case(&b'X')
-    {
-        return Ok(QueryValue::Blob(decode_hex_bytes(
-            &value[2..value.len() - 1],
-        )?));
-    }
-    if value == "Inf" {
-        return Ok(QueryValue::SqlLiteral("1e999".to_string()));
-    }
-    if value == "-Inf" {
-        return Ok(QueryValue::SqlLiteral("-1e999".to_string()));
-    }
-    Ok(QueryValue::SqlLiteral(value.to_string()))
-}
-
-fn parse_quoted_records(
-    stdout: &str,
-    source: QuerySource,
-) -> Result<Vec<QuotedRecord>, DbOperationError> {
-    split_quoted_records(stdout)?
-        .into_iter()
-        .enumerate()
-        .map(|(index, segment)| {
-            let decode_preview_transport = source == QuerySource::Preview && index > 0;
-            split_quoted_fields(&segment.text)?
-                .into_iter()
-                .map(|field| parse_quoted_value(&field, source, decode_preview_transport))
-                .collect::<Result<Vec<_>, _>>()
-                .map(|values| QuotedRecord {
-                    offset: segment.offset,
-                    values,
-                })
-        })
-        .collect()
-}
-
-fn statement_counts_as_select_tag(statement: &str) -> bool {
-    let keyword = first_keyword(statement);
-    keyword.eq_ignore_ascii_case("SELECT") || keyword.eq_ignore_ascii_case("WITH")
-}
-
-fn quoted_to_query_result(
-    query: &str,
-    stdout: &str,
-    source: QuerySource,
-    execution_time_ms: u64,
-) -> Result<QueryResult, DbOperationError> {
-    let stdout = stdout.trim();
-    if stdout.is_empty() {
-        return Ok(QueryResult::success(
-            query.to_string(),
-            Vec::new(),
-            Vec::new(),
-            execution_time_ms,
-            source,
-        ));
-    }
-
-    let mut records = parse_quoted_records(stdout, source)?;
-    let Some(header) = records.first() else {
-        return Ok(QueryResult::success(
-            query.to_string(),
-            Vec::new(),
-            Vec::new(),
-            execution_time_ms,
-            source,
-        ));
-    };
-    let columns = header
-        .values
-        .iter()
-        .map(QueryValue::display_value)
-        .collect();
-    let values = records.drain(1..).map(|record| record.values).collect();
-    Ok(QueryResult::success_with_values(
-        query.to_string(),
-        columns,
-        values,
-        execution_time_ms,
-        source,
-    ))
-}
-
-fn last_sqlite_result_set(stdout: &str, marker: &str) -> Result<Option<String>, DbOperationError> {
-    let (stmt_col, marker_col) = sqlite_result_probe_columns(marker);
-    let raw_records = split_quoted_records(stdout)?;
-    let records = parse_quoted_records(stdout, QuerySource::Adhoc)?;
-
-    let mut last_result = None;
-    let mut result_start = 0;
-    let mut index = 0;
-    while index < records.len() {
-        let record = &records[index];
-        if record.values.len() == 2
-            && record.values[0].as_str() == Some(stmt_col.as_str())
-            && record.values[1].as_str() == Some(marker_col.as_str())
-        {
-            let value = records.get(index + 1).ok_or_else(|| {
-                DbOperationError::CommandTagParseFailed(
-                    "missing SQLite result marker row".to_string(),
-                )
-            })?;
-            let marker_value = value
-                .values
-                .get(1)
-                .and_then(QueryValue::as_str)
-                .ok_or_else(|| {
-                    DbOperationError::CommandTagParseFailed(
-                        "invalid SQLite result marker".to_string(),
-                    )
-                })?;
-            if marker_value != marker {
-                return Err(DbOperationError::CommandTagParseFailed(
-                    "mismatched SQLite result marker".to_string(),
-                ));
-            }
-            last_result = Some(
-                raw_records[result_start..index]
-                    .iter()
-                    .map(|segment| segment.text.as_str())
-                    .collect::<Vec<_>>()
-                    .join("\n"),
-            );
-            index += 2;
-            result_start = index;
-        } else {
-            index += 1;
-        }
-    }
-
-    Ok(last_result)
-}
-
-fn strip_sqlite_probes(
-    stdout: &str,
-    marker: &str,
-) -> Result<(String, HashMap<usize, usize>), DbOperationError> {
-    if stdout.trim().is_empty() {
-        return Ok((String::new(), HashMap::new()));
-    }
-
-    let (stmt_col, changes_col) = sqlite_probe_columns(marker);
-    let raw_records = split_quoted_records(stdout)?;
-    let records = parse_quoted_records(stdout, QuerySource::Adhoc)?;
-
-    let mut changes = HashMap::new();
-    let mut kept = Vec::new();
-    let mut removed_probe = false;
-    let mut index = 0;
-    while index < records.len() {
-        let record = &records[index];
-        if record.values.len() == 2
-            && record.values[0].as_str() == Some(stmt_col.as_str())
-            && record.values[1].as_str() == Some(changes_col.as_str())
-        {
-            removed_probe = true;
-            let value = records.get(index + 1).ok_or_else(|| {
-                DbOperationError::CommandTagParseFailed(
-                    "missing SQLite statement probe row".to_string(),
-                )
-            })?;
-            let stmt_index = value
-                .values
-                .first()
-                .and_then(QueryValue::as_str)
-                .and_then(|raw| raw.parse::<usize>().ok())
-                .ok_or_else(|| {
-                    DbOperationError::CommandTagParseFailed(
-                        "invalid SQLite statement probe index".to_string(),
-                    )
-                })?;
-            let affected_rows = value
-                .values
-                .get(1)
-                .and_then(QueryValue::as_str)
-                .and_then(|raw| raw.parse::<usize>().ok())
-                .ok_or_else(|| {
-                    DbOperationError::CommandTagParseFailed(
-                        "invalid SQLite statement probe changes".to_string(),
-                    )
-                })?;
-            changes.insert(stmt_index, affected_rows);
-            index += 2;
-        } else {
-            kept.push(raw_records[index].text.clone());
-            index += 1;
-        }
-    }
-
-    if !removed_probe {
-        return Ok((stdout.to_string(), changes));
-    }
-
-    Ok((kept.join("\n"), changes))
-}
-
-fn first_csv_cell(stdout: &str) -> Result<String, DbOperationError> {
-    let mut reader = csv::ReaderBuilder::new()
-        .has_headers(true)
-        .from_reader(stdout.trim().as_bytes());
-    let mut records = reader.records();
-    let record = records
-        .next()
-        .transpose()?
-        .ok_or_else(|| DbOperationError::EmptyResponse(stdout.to_string()))?;
-    record
-        .get(0)
-        .map(ToString::to_string)
-        .ok_or_else(|| DbOperationError::EmptyResponse(stdout.to_string()))
-}
-
-fn last_csv_cell(stdout: &str) -> Result<String, DbOperationError> {
-    let line = stdout
-        .lines()
-        .rev()
-        .find(|line| !line.trim().is_empty())
-        .ok_or_else(|| DbOperationError::EmptyResponse(stdout.to_string()))?;
-    let mut reader = csv::ReaderBuilder::new()
-        .has_headers(false)
-        .from_reader(line.as_bytes());
-    let mut records = reader.records();
-    let record = records
-        .next()
-        .transpose()?
-        .ok_or_else(|| DbOperationError::EmptyResponse(stdout.to_string()))?;
-    record
-        .get(0)
-        .map(ToString::to_string)
-        .ok_or_else(|| DbOperationError::EmptyResponse(stdout.to_string()))
-}
-
-fn parse_affected_rows(stdout: &str) -> Result<usize, DbOperationError> {
-    last_csv_cell(stdout)
-        .map_err(|error| match error {
-            DbOperationError::EmptyResponse(_) => {
-                DbOperationError::CommandTagParseFailed(stdout.to_string())
-            }
-            other => other,
-        })?
-        .parse::<usize>()
-        .map_err(|error| DbOperationError::CommandTagParseFailed(error.to_string()))
-}
-
-fn parse_count_result(stdout: &str) -> Result<usize, DbOperationError> {
-    first_csv_cell(stdout)
-        .map_err(|error| match error {
-            DbOperationError::EmptyResponse(_) => {
-                DbOperationError::QueryFailed("Failed to parse COUNT result".to_string())
-            }
-            other => other,
-        })?
-        .parse::<usize>()
-        .map_err(|error| {
-            DbOperationError::QueryFailed(format!("Failed to parse COUNT result: {error}"))
-        })
-}
-
-fn ddl_tag(query: &str) -> Option<CommandTag> {
-    let object = second_keyword(query)
-        .unwrap_or("OBJECT")
-        .to_ascii_uppercase();
-    let keyword = first_keyword(query);
-    if keyword.eq_ignore_ascii_case("CREATE") {
-        Some(CommandTag::Create(object))
-    } else if keyword.eq_ignore_ascii_case("DROP") {
-        Some(CommandTag::Drop(object))
-    } else if keyword.eq_ignore_ascii_case("ALTER") {
-        Some(CommandTag::Alter(object))
-    } else {
-        None
-    }
-}
-
-fn transaction_control_tag(query: &str) -> Option<CommandTag> {
-    match first_keyword(query).to_ascii_uppercase().as_str() {
-        "BEGIN" => Some(CommandTag::Begin),
-        "COMMIT" | "END" => Some(CommandTag::Commit),
-        "ROLLBACK" if is_rollback_to(query) => Some(CommandTag::Other(format!(
-            "ROLLBACK TO {}",
-            rollback_to_target(query).unwrap_or("")
-        ))),
-        "ROLLBACK" => Some(CommandTag::Rollback),
-        "SAVEPOINT" => Some(CommandTag::Other(format!(
-            "SAVEPOINT {}",
-            second_keyword(query).unwrap_or("")
-        ))),
-        "RELEASE" => Some(CommandTag::Other(format!(
-            "RELEASE {}",
-            second_keyword(query).unwrap_or("")
-        ))),
-        _ => None,
-    }
-}
-
-fn dml_tag(query: &str, affected_rows: usize) -> Option<CommandTag> {
-    let affected_rows = affected_rows as u64;
-    match dml_keyword(query) {
-        Some("INSERT") => Some(CommandTag::Insert(affected_rows)),
-        Some("UPDATE") => Some(CommandTag::Update(affected_rows)),
-        Some("DELETE") => Some(CommandTag::Delete(affected_rows)),
-        _ => None,
-    }
-}
-
-fn sqlite_side_effect_tag(query: &str) -> Option<CommandTag> {
-    let keyword = first_keyword(query).to_ascii_uppercase();
-    match keyword.as_str() {
-        "ANALYZE" | "ATTACH" | "DETACH" | "REINDEX" | "VACUUM" => Some(CommandTag::Other(keyword)),
-        _ => None,
-    }
-}
-
-fn sqlite_statement_tags(statements: &[&str], changes: &HashMap<usize, usize>) -> Vec<CommandTag> {
-    statements
-        .iter()
-        .enumerate()
-        .filter_map(|(index, statement)| {
-            dml_tag(statement, *changes.get(&index).unwrap_or(&0))
-                .or_else(|| ddl_tag(statement))
-                .or_else(|| transaction_control_tag(statement))
-                .or_else(|| sqlite_side_effect_tag(statement))
-        })
-        .collect()
-}
-
-fn discard_rolled_back(tags: &[CommandTag]) -> Vec<CommandTag> {
-    let mut effective = Vec::new();
-    let mut frames: Vec<(Option<String>, Vec<CommandTag>)> = Vec::new();
-
-    for tag in tags {
-        match tag {
-            CommandTag::Begin => frames.push((None, Vec::new())),
-            CommandTag::Other(raw) if raw == "SAVEPOINT" || raw.starts_with("SAVEPOINT ") => {
-                frames.push((tag_name(raw, "SAVEPOINT"), Vec::new()));
-            }
-            CommandTag::Other(raw) if raw == "RELEASE" || raw.starts_with("RELEASE ") => {
-                if let Some(index) = savepoint_frame_index(&frames, tag_name(raw, "RELEASE")) {
-                    let mut merged = Vec::new();
-                    for (_, frame) in frames.drain(index..) {
-                        merged.extend(frame);
-                    }
-                    if let Some((_, parent)) = frames.last_mut() {
-                        parent.extend(merged);
-                    } else {
-                        effective.extend(merged);
-                    }
-                }
-            }
-            CommandTag::Other(raw) if raw == "ROLLBACK TO" || raw.starts_with("ROLLBACK TO ") => {
-                if let Some(index) = savepoint_frame_index(&frames, tag_name(raw, "ROLLBACK TO")) {
-                    frames.truncate(index + 1);
-                    if let Some((_, frame)) = frames.last_mut() {
-                        frame.clear();
-                    }
-                }
-            }
-            CommandTag::Rollback => {
-                frames.clear();
-            }
-            CommandTag::Commit => {
-                for (_, frame) in frames.drain(..) {
-                    effective.extend(frame);
-                }
-            }
-            _ => {
-                if let Some((_, frame)) = frames.last_mut() {
-                    frame.push(tag.clone());
-                } else {
-                    effective.push(tag.clone());
-                }
-            }
-        }
-    }
-
-    effective
-}
-
-fn tag_name(raw: &str, prefix: &str) -> Option<String> {
-    raw.strip_prefix(prefix)
-        .map(str::trim)
-        .filter(|name| !name.is_empty())
-        .map(str::to_ascii_uppercase)
-}
-
-fn savepoint_frame_index(
-    frames: &[(Option<String>, Vec<CommandTag>)],
-    name: Option<String>,
-) -> Option<usize> {
-    frames
-        .iter()
-        .enumerate()
-        .rev()
-        .find_map(|(index, (frame_name, _))| {
-            if frame_name.is_none() && index == 0 {
-                return None;
-            }
-            if name
-                .as_ref()
-                .is_none_or(|name| frame_name.as_ref() == Some(name))
-            {
-                Some(index)
-            } else {
-                None
-            }
-        })
-}
-
-fn aggregate_sqlite_command_tag(tags: &[CommandTag]) -> Option<CommandTag> {
-    let effective = discard_rolled_back(tags);
-    if let Some(tag) = effective.iter().find(|tag| tag.is_schema_modifying()) {
-        return Some(tag.clone());
-    }
-    if let Some(tag) = effective.iter().rev().find(|tag| tag.needs_refresh()) {
-        return Some(tag.clone());
-    }
-    if tags.iter().any(CommandTag::needs_refresh) {
-        return Some(CommandTag::Rollback);
-    }
-    tags.last().cloned()
-}
-
-fn command_tag_result(
-    query: &str,
-    tag: CommandTag,
-    elapsed: u64,
-    source: QuerySource,
-) -> QueryResult {
-    QueryResult::success(query.to_string(), Vec::new(), Vec::new(), elapsed, source)
-        .with_row_count(tag.affected_rows().unwrap_or(0) as usize)
-        .with_command_tag(tag)
 }
 
 fn parse_fk_action(action: &str) -> Result<FkAction, DbOperationError> {
@@ -2257,365 +884,6 @@ mod tests {
     use crate::domain::{CommandTag, DatabaseType, QuerySource};
 
     use super::*;
-
-    #[test]
-    fn split_sqlite_statements_ignores_semicolons_in_literals_and_comments() {
-        let statements = try_split_sqlite_statements(
-            "INSERT INTO logs(message) VALUES ('a;b'); -- ; ignored\nSELECT ';' AS value;",
-        )
-        .unwrap();
-
-        assert_eq!(
-            statements,
-            vec![
-                "INSERT INTO logs(message) VALUES ('a;b')",
-                "-- ; ignored\nSELECT ';' AS value"
-            ]
-        );
-    }
-
-    #[test]
-    fn split_sqlite_statements_keeps_create_trigger_body_together() {
-        let trigger = "\
-CREATE TRIGGER agent_messages_fts_ai AFTER INSERT ON agent_messages BEGIN
-    INSERT INTO agent_messages_fts(rowid, role, content)
-    VALUES (new.id, new.role, new.content);
-END";
-        let sql = format!("{trigger}; SELECT 1 AS value;");
-
-        let statements = try_split_sqlite_statements(&sql).unwrap();
-
-        assert_eq!(statements.len(), 2);
-        assert_eq!(statements[0], trigger);
-        assert_eq!(statements[1], "SELECT 1 AS value");
-    }
-
-    #[test]
-    fn split_sqlite_statements_keeps_create_trigger_with_dotted_end_reference() {
-        let trigger = "\
-CREATE TRIGGER sync_end AFTER UPDATE ON events BEGIN
-    UPDATE counters SET end_value = new.end WHERE id = new.id;
-    INSERT INTO audit(event_id, end_value) VALUES (new.id, new.end);
-END";
-        let sql = format!("{trigger}; SELECT 1 AS value;");
-
-        let statements = try_split_sqlite_statements(&sql).unwrap();
-
-        assert_eq!(statements.len(), 2);
-        assert_eq!(statements[0], trigger);
-        assert_eq!(statements[1], "SELECT 1 AS value");
-    }
-
-    #[test]
-    fn split_sqlite_statements_keeps_create_trigger_with_case_end_expression() {
-        let trigger = "\
-CREATE TRIGGER normalize_events AFTER UPDATE ON events BEGIN
-    UPDATE counters
-    SET end_value = CASE WHEN new.end > 0 THEN new.end ELSE old.end END
-    WHERE id = new.id;
-    INSERT INTO audit(event_id) VALUES (new.id);
-END";
-        let sql = format!("{trigger}; SELECT 1 AS value;");
-
-        let statements = try_split_sqlite_statements(&sql).unwrap();
-
-        assert_eq!(statements.len(), 2);
-        assert_eq!(statements[0], trigger);
-        assert_eq!(statements[1], "SELECT 1 AS value");
-    }
-
-    #[test]
-    fn adhoc_execution_query_does_not_insert_probes_when_trigger_references_new_end() {
-        let trigger = "\
-CREATE TRIGGER sync_end AFTER UPDATE ON events BEGIN
-    UPDATE counters SET end_value = new.end WHERE id = new.id;
-    INSERT INTO audit(event_id, end_value) VALUES (new.id, new.end);
-END";
-        let marker = "probe_marker";
-
-        let execution_query = sqlite_adhoc_execution_query(trigger, marker).unwrap();
-
-        assert!(!execution_query.contains(marker));
-        assert_eq!(execution_query, trigger);
-    }
-
-    #[test]
-    fn split_sqlite_statements_rejects_unclosed_create_trigger_body() {
-        let error = try_split_sqlite_statements(
-            "CREATE TRIGGER t AFTER INSERT ON users BEGIN INSERT INTO logs(id) VALUES (1);",
-        )
-        .unwrap_err();
-
-        assert!(matches!(error, DbOperationError::QueryFailed(_)));
-    }
-
-    #[test]
-    fn split_sqlite_statements_rejects_incomplete_create_trigger_without_begin() {
-        let error =
-            try_split_sqlite_statements("CREATE TRIGGER t AFTER INSERT ON users").unwrap_err();
-
-        assert!(matches!(error, DbOperationError::QueryFailed(_)));
-    }
-
-    #[test]
-    fn adhoc_execution_query_does_not_insert_probes_inside_create_trigger() {
-        let trigger = "\
-CREATE TRIGGER agent_messages_fts_ai AFTER INSERT ON agent_messages BEGIN
-    INSERT INTO agent_messages_fts(rowid, role, content)
-    VALUES (new.id, new.role, new.content);
-END";
-        let marker = "probe_marker";
-
-        let execution_query = sqlite_adhoc_execution_query(trigger, marker).unwrap();
-
-        assert!(!execution_query.contains(marker));
-        assert_eq!(execution_query, trigger);
-    }
-
-    #[test]
-    fn append_changes_wraps_multi_statement_write_without_explicit_transaction() {
-        let query = "INSERT INTO users(id) VALUES (1); INSERT INTO users(id) VALUES (2);";
-
-        let wrapped = append_changes_query(query).unwrap();
-
-        assert_eq!(
-            wrapped,
-            "BEGIN;\nINSERT INTO users(id) VALUES (1); INSERT INTO users(id) VALUES (2)\n;\nCOMMIT\n;\nSELECT changes() AS affected_rows;"
-        );
-    }
-
-    #[test]
-    fn append_changes_wraps_multi_statement_replace_without_explicit_transaction() {
-        let query = "REPLACE INTO users(id) VALUES (1); SELECT * FROM missing";
-
-        let wrapped = append_changes_query(query).unwrap();
-
-        assert_eq!(
-            wrapped,
-            "BEGIN;\nREPLACE INTO users(id) VALUES (1); SELECT * FROM missing\n;\nCOMMIT\n;\nSELECT changes() AS affected_rows;"
-        );
-    }
-
-    #[test]
-    fn append_changes_wraps_multi_statement_with_write_without_explicit_transaction() {
-        let query = "WITH payload(id) AS (VALUES (1)) INSERT INTO users(id) SELECT id FROM payload; SELECT * FROM missing";
-
-        let wrapped = append_changes_query(query).unwrap();
-
-        assert_eq!(
-            wrapped,
-            "BEGIN;\nWITH payload(id) AS (VALUES (1)) INSERT INTO users(id) SELECT id FROM payload; SELECT * FROM missing\n;\nCOMMIT\n;\nSELECT changes() AS affected_rows;"
-        );
-    }
-
-    #[test]
-    fn append_changes_keeps_explicit_begin_commit_transaction_control() {
-        let query = "BEGIN; INSERT INTO users(id) VALUES (1); COMMIT";
-
-        let wrapped = append_changes_query(query).unwrap();
-
-        assert_eq!(
-            wrapped,
-            "BEGIN; INSERT INTO users(id) VALUES (1); COMMIT\n;\nSELECT changes() AS affected_rows;"
-        );
-    }
-
-    #[test]
-    fn append_changes_keeps_explicit_begin_end_transaction_control() {
-        let query = "BEGIN; INSERT INTO users(id) VALUES (1); END";
-
-        let wrapped = append_changes_query(query).unwrap();
-
-        assert_eq!(
-            wrapped,
-            "BEGIN; INSERT INTO users(id) VALUES (1); END\n;\nSELECT changes() AS affected_rows;"
-        );
-    }
-
-    mod discard_rolled_back_policy {
-        use super::*;
-
-        fn sp(name: &str) -> CommandTag {
-            CommandTag::Other(format!("SAVEPOINT {name}"))
-        }
-
-        fn rollback_to(name: &str) -> CommandTag {
-            CommandTag::Other(format!("ROLLBACK TO {name}"))
-        }
-
-        fn release(name: &str) -> CommandTag {
-            CommandTag::Other(format!("RELEASE {name}"))
-        }
-
-        #[test]
-        fn top_level_savepoint_rollback_to_discards_inner_dml() {
-            let tags = vec![
-                sp("sp"),
-                CommandTag::Insert(1),
-                rollback_to("sp"),
-                CommandTag::Insert(2),
-                release("sp"),
-            ];
-
-            assert_eq!(discard_rolled_back(&tags), vec![CommandTag::Insert(2)]);
-        }
-
-        #[test]
-        fn top_level_savepoint_full_rollback_discards_all_dml() {
-            let tags = vec![sp("sp"), CommandTag::Insert(1), CommandTag::Rollback];
-
-            assert!(discard_rolled_back(&tags).is_empty());
-        }
-
-        #[test]
-        fn unclosed_begin_discards_frame_from_effective() {
-            let tags = vec![CommandTag::Begin, CommandTag::Update(1)];
-
-            assert!(discard_rolled_back(&tags).is_empty());
-        }
-
-        #[test]
-        fn unclosed_top_level_savepoint_discards_frame_from_effective() {
-            let tags = vec![sp("sp"), CommandTag::Insert(1)];
-
-            assert!(discard_rolled_back(&tags).is_empty());
-        }
-
-        #[test]
-        fn unclosed_top_level_savepoint_aggregates_to_rollback() {
-            let tags = vec![sp("sp"), CommandTag::Insert(1)];
-
-            assert_eq!(
-                aggregate_sqlite_command_tag(&tags),
-                Some(CommandTag::Rollback)
-            );
-        }
-
-        #[test]
-        fn rollback_to_quoted_savepoint_is_not_full_rollback() {
-            let tags = sqlite_statement_tags(
-                &[
-                    "SAVEPOINT sp",
-                    "INSERT INTO users(id) VALUES (1)",
-                    "ROLLBACK TO \"sp\"",
-                ],
-                &HashMap::from([(1, 1)]),
-            );
-
-            assert_eq!(
-                tags,
-                vec![
-                    CommandTag::Other("SAVEPOINT sp".to_string()),
-                    CommandTag::Insert(1),
-                    CommandTag::Other("ROLLBACK TO ".to_string()),
-                ]
-            );
-            assert!(discard_rolled_back(&tags).is_empty());
-        }
-
-        #[test]
-        fn begin_savepoint_release_still_merges_nested_frame() {
-            let tags = vec![
-                CommandTag::Begin,
-                sp("inner"),
-                CommandTag::Insert(1),
-                release("inner"),
-                CommandTag::Commit,
-            ];
-
-            assert_eq!(discard_rolled_back(&tags), vec![CommandTag::Insert(1)]);
-        }
-
-        #[test]
-        fn rollback_transaction_to_savepoint_discards_inner_dml() {
-            let tags = sqlite_statement_tags(
-                &[
-                    "SAVEPOINT sp",
-                    "INSERT INTO users(id) VALUES (1)",
-                    "INSERT INTO users(id) VALUES (2)",
-                    "ROLLBACK TRANSACTION TO SAVEPOINT sp",
-                    "INSERT INTO users(id) VALUES (3)",
-                    "RELEASE sp",
-                ],
-                &HashMap::from([(1, 1), (2, 1), (4, 1)]),
-            );
-
-            assert_eq!(discard_rolled_back(&tags), vec![CommandTag::Insert(1)]);
-        }
-    }
-
-    mod wrap_mode {
-        use super::*;
-
-        #[test]
-        fn autocommit_multi_write_uses_begin_commit() {
-            assert_eq!(
-                sqlite_wrap_mode(
-                    "INSERT INTO users(id) VALUES (1); INSERT INTO users(id) VALUES (2)"
-                )
-                .unwrap(),
-                SqliteWrapMode::BeginCommit
-            );
-        }
-
-        #[test]
-        fn explicit_begin_is_not_wrapped() {
-            assert_eq!(
-                sqlite_wrap_mode("BEGIN; INSERT INTO users(id) VALUES (1); COMMIT").unwrap(),
-                SqliteWrapMode::None
-            );
-        }
-
-        #[test]
-        fn top_level_savepoint_multi_write_is_not_wrapped() {
-            assert_eq!(
-                sqlite_wrap_mode(
-                    "SAVEPOINT user_sp; INSERT INTO users(id) VALUES (1); INSERT INTO users(id) VALUES (2)"
-                )
-                .unwrap(),
-                SqliteWrapMode::None
-            );
-        }
-
-        #[test]
-        fn mid_batch_savepoint_is_not_wrapped() {
-            assert_eq!(
-                sqlite_wrap_mode(
-                    "INSERT INTO users(id) VALUES (1); SAVEPOINT sp; INSERT INTO users(id) VALUES (2)"
-                )
-                .unwrap(),
-                SqliteWrapMode::None
-            );
-        }
-    }
-
-    #[test]
-    fn sqlite_side_effect_statements_emit_refresh_tags() {
-        let changes = HashMap::new();
-
-        let tags = sqlite_statement_tags(
-            &[
-                "ANALYZE users",
-                "ATTACH DATABASE 'other.db' AS other",
-                "DETACH DATABASE other",
-                "REINDEX users_name_idx",
-                "VACUUM",
-            ],
-            &changes,
-        );
-
-        assert_eq!(
-            tags,
-            vec![
-                CommandTag::Other("ANALYZE".to_string()),
-                CommandTag::Other("ATTACH".to_string()),
-                CommandTag::Other("DETACH".to_string()),
-                CommandTag::Other("REINDEX".to_string()),
-                CommandTag::Other("VACUUM".to_string()),
-            ]
-        );
-        assert!(tags.iter().all(CommandTag::needs_refresh));
-    }
 
     #[test]
     fn index_key_column_names_preserves_expression_and_unknown_key_columns() {
@@ -2842,259 +1110,6 @@ END";
                 result.value_at(0, 2),
                 Some(&QueryValue::Text("ALICE".to_string()))
             );
-        }
-    }
-
-    mod parsing {
-        use super::*;
-
-        #[test]
-        fn quoted_to_query_result_preserves_newline_for_single_statement() {
-            let quoted = "'body','marker'\n'line 1\nline 2','ok'\n";
-
-            let result = quoted_to_query_result(
-                "SELECT body, marker FROM notes",
-                quoted,
-                QuerySource::Adhoc,
-                1,
-            )
-            .unwrap();
-
-            assert_eq!(result.columns, vec!["body", "marker"]);
-            assert_eq!(
-                result.rows(),
-                vec![vec!["line 1\nline 2".to_string(), "ok".to_string()]]
-            );
-        }
-
-        #[test]
-        fn last_sqlite_result_set_uses_marker_boundaries() {
-            let marker = "probe";
-            let sqlite_output_with_ignored_first_result_set = "'ignored'\n1\n'probe_result_stmt','probe_result_marker'\n0,'probe'\n'body','marker'\n'line 1\nline 2','ok'\n'probe_result_stmt','probe_result_marker'\n1,'probe'\n";
-
-            let quoted =
-                last_sqlite_result_set(sqlite_output_with_ignored_first_result_set, marker)
-                    .unwrap()
-                    .unwrap();
-            let result = quoted_to_query_result(
-                "SELECT 1 AS ignored; SELECT body, marker FROM notes",
-                &quoted,
-                QuerySource::Adhoc,
-                1,
-            )
-            .unwrap();
-
-            assert_eq!(result.columns, vec!["body", "marker"]);
-            assert_eq!(
-                result.rows(),
-                vec![vec!["line 1\nline 2".to_string(), "ok".to_string()]]
-            );
-        }
-
-        #[test]
-        fn quoted_to_query_result_preserves_sqlite_value_kinds() {
-            let quoted = "'a','b','c','d'\nNULL,'','NULL',X'00FF41'\n";
-
-            let result =
-                quoted_to_query_result("SELECT a, b, c, d FROM t", quoted, QuerySource::Adhoc, 1)
-                    .unwrap();
-
-            assert_eq!(
-                result.rows(),
-                vec![vec![
-                    "NULL".to_string(),
-                    String::new(),
-                    "NULL".to_string(),
-                    "BLOB (3 bytes) 00 FF 41".to_string()
-                ]]
-            );
-            assert!(matches!(result.value_at(0, 0), Some(QueryValue::Null)));
-            assert_eq!(
-                result.value_at(0, 1),
-                Some(&QueryValue::Text(String::new()))
-            );
-            assert_eq!(
-                result.value_at(0, 2),
-                Some(&QueryValue::Text("NULL".to_string()))
-            );
-            assert_eq!(
-                result.value_at(0, 3),
-                Some(&QueryValue::Blob(vec![0, 255, 65]))
-            );
-        }
-
-        #[test]
-        fn quoted_to_query_result_normalizes_infinite_numeric_literals() {
-            let quoted = "'pos','neg'\nInf,-Inf\n";
-
-            let result =
-                quoted_to_query_result("SELECT 1e999, -1e999", quoted, QuerySource::Adhoc, 1)
-                    .unwrap();
-
-            assert_eq!(
-                result.value_at(0, 0),
-                Some(&QueryValue::SqlLiteral("1e999".to_string()))
-            );
-            assert_eq!(
-                result.value_at(0, 1),
-                Some(&QueryValue::SqlLiteral("-1e999".to_string()))
-            );
-        }
-
-        #[test]
-        fn quoted_to_query_result_rejects_unterminated_quote() {
-            let result = quoted_to_query_result(
-                "SELECT body FROM notes",
-                "'body'\n'unclosed\nnext",
-                QuerySource::Adhoc,
-                1,
-            );
-
-            assert!(matches!(
-                result,
-                Err(DbOperationError::MetadataParseFailed(message))
-                    if message == "unterminated SQLite quoted output"
-            ));
-        }
-
-        #[test]
-        fn parse_unistr_inner_sql_escapes_does_not_decode_unicode_sequences() {
-            assert_eq!(
-                parse_unistr_inner_sql_escapes("unistr('\\u0001\\u0001')").unwrap(),
-                "\\u0001\\u0001"
-            );
-            assert_eq!(
-                parse_unistr_inner_sql_escapes("unistr('\\u0001O''Reilly')").unwrap(),
-                "\\u0001O'Reilly"
-            );
-        }
-
-        #[test]
-        fn decode_preview_transport_unistr_decodes_hex_payload() {
-            assert_eq!(
-                decode_preview_transport_unistr("unistr('\\u0001SABIQL_HEX:61006263')")
-                    .unwrap()
-                    .as_deref(),
-                Some("a\0bc")
-            );
-            assert_eq!(
-                decode_preview_transport_unistr("unistr('\\u0001SABIQL_HEX:0101')")
-                    .unwrap()
-                    .as_deref(),
-                Some("\x01\x01")
-            );
-            assert_eq!(
-                decode_preview_transport_unistr("unistr('\\u0001SABIQL_HEX:015C7530303031')")
-                    .unwrap()
-                    .as_deref(),
-                Some("\x01\\u0001")
-            );
-        }
-
-        #[test]
-        fn parse_quoted_value_keeps_unrecoverable_adhoc_unistr_as_sql_literal() {
-            assert_eq!(
-                parse_quoted_value("unistr('\\u0001\\u0001')", QuerySource::Adhoc, true).unwrap(),
-                QueryValue::SqlLiteral("unistr('\\u0001\\u0001')".to_string())
-            );
-            assert_eq!(
-                parse_quoted_value("unistr('\\u0001O''Reilly')", QuerySource::Adhoc, true).unwrap(),
-                QueryValue::SqlLiteral("unistr('\\u0001O''Reilly')".to_string())
-            );
-        }
-
-        #[test]
-        fn parse_quoted_value_decodes_preview_transport_unistr() {
-            let value = parse_quoted_value(
-                "unistr('\\u0001SABIQL_HEX:61006263')",
-                QuerySource::Preview,
-                true,
-            )
-            .unwrap();
-
-            assert_eq!(value, QueryValue::Text("a\0bc".to_string()));
-        }
-
-        #[test]
-        fn parse_quoted_value_decodes_preview_transport_plain_quoted() {
-            let sentinel = sql::sqlite_nul_text_sentinel();
-            let quoted = format!("'{sentinel}68656C6C6F'");
-            let value = parse_quoted_value(&quoted, QuerySource::Preview, true).unwrap();
-
-            assert_eq!(value, QueryValue::Text("hello".to_string()));
-        }
-
-        #[test]
-        fn parse_quoted_value_keeps_plain_quoted_transport_as_text_for_adhoc() {
-            let sentinel = sql::sqlite_nul_text_sentinel();
-            let transport = format!("{sentinel}68656C6C6F");
-            let quoted = format!("'{transport}'");
-            let value = parse_quoted_value(&quoted, QuerySource::Adhoc, true).unwrap();
-
-            assert_eq!(value, QueryValue::Text(transport));
-        }
-
-        #[test]
-        fn parse_quoted_value_skips_preview_transport_decode_for_column_names() {
-            let sentinel = sql::sqlite_nul_text_sentinel();
-            let transport = format!("{sentinel}68656C6C6F");
-            let quoted = format!("'{transport}'");
-            let value = parse_quoted_value(&quoted, QuerySource::Preview, false).unwrap();
-
-            assert_eq!(value, QueryValue::Text(transport));
-        }
-
-        #[test]
-        fn quoted_to_query_result_keeps_transport_like_column_name() {
-            let sentinel = sql::sqlite_nul_text_sentinel();
-            let column = format!("{sentinel}4142");
-            let data = format!("'{sentinel}68656C6C6F'");
-            let quoted = format!("'{column}'\n{data}\n");
-            let result =
-                quoted_to_query_result("SELECT 1", &quoted, QuerySource::Preview, 1).unwrap();
-
-            assert_eq!(result.columns, vec![column]);
-            assert_eq!(
-                result.value_at(0, 0),
-                Some(&QueryValue::Text("hello".to_string()))
-            );
-        }
-
-        #[test]
-        fn quoted_to_query_result_keeps_unrecoverable_adhoc_unistr_as_sql_literal() {
-            let quoted = "'value'\nunistr('\\u0001\\u0001')\n";
-
-            let result =
-                quoted_to_query_result("SELECT char(1) || char(1)", quoted, QuerySource::Adhoc, 1)
-                    .unwrap();
-
-            assert_eq!(
-                result.value_at(0, 0),
-                Some(&QueryValue::SqlLiteral(
-                    "unistr('\\u0001\\u0001')".to_string()
-                ))
-            );
-        }
-
-        #[test]
-        fn parse_affected_rows_reads_trailing_changes_cell() {
-            assert_eq!(parse_affected_rows("changes()\n3\n").unwrap(), 3);
-        }
-
-        #[test]
-        fn strip_sqlite_probes_removes_probe_result_sets() {
-            let marker = "probe";
-            let stdout = "'id','name'\n1,'Alice'\n'probe_stmt','probe_changes'\n0,2\n'value'\n42\n";
-
-            let (filtered, changes) = strip_sqlite_probes(stdout, marker).unwrap();
-
-            assert_eq!(changes.get(&0), Some(&2));
-            assert_eq!(filtered, "'id','name'\n1,'Alice'\n'value'\n42");
-        }
-
-        #[test]
-        fn parse_count_result_reads_first_result_cell() {
-            assert_eq!(parse_count_result("COUNT(*)\n42\n").unwrap(), 42);
         }
     }
 
@@ -3350,7 +1365,7 @@ END";
                 .await
                 .unwrap();
 
-            assert_eq!(result.rows(), vec![vec![cli::BUSY_TIMEOUT_MS.to_string()]]);
+            assert_eq!(result.rows(), vec![vec![BUSY_TIMEOUT_MS.to_string()]]);
         }
 
         #[tokio::test]
@@ -4089,25 +2104,6 @@ END";
             assert_eq!(count, 3);
         }
 
-        #[test]
-        fn export_guard_rejects_non_rerunnable_sql() {
-            for sql in [
-                "SELECT 1; SELECT 2",
-                "WITH payload(id) AS (VALUES (1)) INSERT INTO users(id) SELECT id FROM payload",
-                "PRAGMA foreign_keys=OFF",
-                "PRAGMA wal_checkpoint(TRUNCATE)",
-            ] {
-                assert!(!is_sqlite_rerunnable_export_query(sql).unwrap(), "{sql}");
-            }
-        }
-
-        #[test]
-        fn export_guard_allows_read_only_sql() {
-            for sql in ["SELECT 1", "PRAGMA table_info(users)"] {
-                assert!(is_sqlite_rerunnable_export_query(sql).unwrap(), "{sql}");
-            }
-        }
-
         #[tokio::test]
         async fn export_to_csv_writes_rows_and_returns_row_count() {
             let (dir, dsn) = make_sqlite_db(
@@ -4367,33 +2363,6 @@ END";
 
             assert!(light.row_count_estimate.is_none());
             assert_eq!(full.row_count_estimate, Some(3));
-        }
-
-        #[tokio::test]
-        async fn loads_table_bound_triggers() {
-            let (_dir, dsn) = make_sqlite_db(
-                r"
-            CREATE TABLE users(id INTEGER PRIMARY KEY, name TEXT);
-            CREATE TABLE audit(user_id INTEGER);
-            CREATE TRIGGER users_audit AFTER INSERT ON users BEGIN
-                INSERT INTO audit(user_id) VALUES (new.id);
-            END;
-            ",
-            );
-            let adapter = SqliteAdapter::new();
-
-            let detail = adapter
-                .fetch_table_detail(&dsn, "main", "users")
-                .await
-                .unwrap();
-
-            assert_eq!(detail.triggers.len(), 1);
-            let trigger = &detail.triggers[0];
-            assert_eq!(trigger.name, "users_audit");
-            assert_eq!(trigger.timing, TriggerTiming::After);
-            assert_eq!(trigger.events, vec![TriggerEvent::Insert]);
-            assert!(trigger.function_name.contains("CREATE TRIGGER users_audit"));
-            assert!(!trigger.security_definer);
         }
 
         #[tokio::test]
@@ -5039,113 +3008,6 @@ END";
                     "idx=idx_users_name:name:false:false:false:false:true:false:true:CREATE INDEX idx_users_name ON users(name COLLATE NOCASE)"
                 )
             );
-        }
-
-        #[tokio::test]
-        async fn trigger_change_updates_signature() {
-            let setup = r"
-            CREATE TABLE users(id INTEGER PRIMARY KEY);
-            CREATE TABLE audit(user_id INTEGER);
-            ";
-            let trigger = r"
-            CREATE TRIGGER users_audit AFTER INSERT ON users BEGIN
-                INSERT INTO audit(user_id) VALUES (new.id);
-            END;
-            ";
-            let (_dir, dsn) = make_sqlite_db(setup);
-            let adapter = SqliteAdapter::new();
-
-            let before = adapter.fetch_table_signatures(&dsn).await.unwrap();
-            let before_signature = before
-                .iter()
-                .find(|signature| signature.name == "users")
-                .unwrap()
-                .signature
-                .clone();
-
-            adapter.execute_adhoc(&dsn, trigger, false).await.unwrap();
-
-            let after = adapter.fetch_table_signatures(&dsn).await.unwrap();
-            let after_signature = &after
-                .iter()
-                .find(|signature| signature.name == "users")
-                .unwrap()
-                .signature;
-
-            assert_ne!(before_signature, after_signature.as_str());
-            assert!(
-                after_signature.contains("trg=users_audit:AFTER:INSERT:CREATE TRIGGER users_audit")
-            );
-        }
-    }
-
-    mod trigger_parsing {
-        use super::*;
-
-        #[test]
-        fn parses_after_insert_trigger() {
-            let sql = "CREATE TRIGGER users_audit AFTER INSERT ON users BEGIN SELECT 1; END";
-            let trigger = parse_sqlite_trigger("users_audit", sql).unwrap();
-
-            assert_eq!(trigger.name, "users_audit");
-            assert_eq!(trigger.timing, TriggerTiming::After);
-            assert_eq!(trigger.events, vec![TriggerEvent::Insert]);
-            assert_eq!(trigger.function_name, sql);
-            assert!(!trigger.security_definer);
-        }
-
-        #[test]
-        fn parses_before_update_of_columns() {
-            let sql =
-                "CREATE TRIGGER users_guard BEFORE UPDATE OF name ON users BEGIN SELECT 1; END";
-            let trigger = parse_sqlite_trigger("users_guard", sql).unwrap();
-
-            assert_eq!(trigger.timing, TriggerTiming::Before);
-            assert_eq!(trigger.events, vec![TriggerEvent::Update]);
-        }
-
-        #[test]
-        fn parses_instead_of_delete_trigger() {
-            let sql =
-                "CREATE TRIGGER users_view_io INSTEAD OF DELETE ON users_view BEGIN SELECT 1; END";
-            let trigger = parse_sqlite_trigger("users_view_io", sql).unwrap();
-
-            assert_eq!(trigger.timing, TriggerTiming::InsteadOf);
-            assert_eq!(trigger.events, vec![TriggerEvent::Delete]);
-        }
-
-        #[test]
-        fn omitted_timing_defaults_to_before() {
-            let sql = "CREATE TRIGGER users_log INSERT ON users BEGIN SELECT 1; END";
-            let trigger = parse_sqlite_trigger("users_log", sql).unwrap();
-
-            assert_eq!(trigger.timing, TriggerTiming::Before);
-            assert_eq!(trigger.events, vec![TriggerEvent::Insert]);
-        }
-    }
-
-    mod trigger_metadata {
-        use super::*;
-
-        #[tokio::test]
-        async fn table_detail_loads_trigger_without_explicit_timing() {
-            let (_dir, dsn) = make_sqlite_db(
-                r"
-            CREATE TABLE users(id INTEGER PRIMARY KEY);
-            CREATE TRIGGER users_log INSERT ON users BEGIN SELECT 1; END;
-            ",
-            );
-            let adapter = SqliteAdapter::new();
-
-            let detail = adapter
-                .fetch_table_detail(&dsn, "main", "users")
-                .await
-                .unwrap();
-
-            assert_eq!(detail.triggers.len(), 1);
-            assert_eq!(detail.triggers[0].name, "users_log");
-            assert_eq!(detail.triggers[0].timing, TriggerTiming::Before);
-            assert_eq!(detail.triggers[0].events, vec![TriggerEvent::Insert]);
         }
     }
 }
