+++ src/infra/adapters/sqlite/sqlite3/error.rs
