+++ src/infra/adapters/sqlite/sqlite3/mod.rs
@@ -1,0 +1,7 @@
+mod error;
+mod executor;
+pub(super) mod parser;
+
+#[cfg(test)]
+pub(in crate::adapters::sqlite) use executor::BUSY_TIMEOUT_MS;
+pub(super) use executor::SqliteCli;
